Faked out the validation procedure for now

+++ src/main/java/cern/modesti/validation/ValidationService.java
@@ -45,6 +45,12 @@ public class ValidationService {
    */
   @Transactional
   public boolean validateRequest(Request request) {
+
+    // TODO: remove this when the validation procedure is ready
+    if (true) {
+      return true;
+    }
+
     // Delete all points with this request id
     String query = "DELETE FROM DRAFT_POINTS WHERE drp_request_id = ?";
     jdbcTemplate.update(query, request.getRequestId());
